+++ src/template_cpu.h
@@ -9,9 +9,9 @@
 #define TEMPLATE_CPU_H_
 // #define RAND_MAX 10
 
-void randomI(double *data, int size);
-void fprintMat(FILE *fp, double *X, char *name, int numVars, int numSamples);
-void printMat(double *X, char *name, int numVars, int numSamples);
+void randomI(float *data, int size);
+void fprintMat(FILE *fp, float *X, char *name, int numVars, int numSamples);
+void printMat(float *X, char *name, int numVars, int numSamples);
 void fprintMatf(FILE *fp, float *X, char *name, int numVars, int numSamples);
 void printMatf(float *X, char *name, int numVars, int numSamples);
